Add features endpoint
